add fee voucher options and create separate components
